+++ app/src/main/java/dev/timmo/systembridge/activity/EditConnectionActivity.kt
@@ -15,15 +15,18 @@ import androidx.appcompat.app.AppCompatActivity
 import androidx.core.widget.addTextChangedListener
 import com.google.android.material.textfield.TextInputEditText
 import com.google.android.material.textfield.TextInputLayout
-import dev.timmo.systembridge.Constants.CONNECTION_API_KEY
-import dev.timmo.systembridge.Constants.CONNECTION_API_PORT
-import dev.timmo.systembridge.Constants.CONNECTION_HOST
-import dev.timmo.systembridge.Constants.CONNECTION_NAME
-import dev.timmo.systembridge.Constants.CONNECTION_UID
-import dev.timmo.systembridge.Constants.DEFAULT_API_PORT
-import dev.timmo.systembridge.Constants.SETUP_EDIT
+import dev.timmo.systembridge.shared.Constants.CONNECTION_API_KEY
+import dev.timmo.systembridge.shared.Constants.CONNECTION_API_PORT
+import dev.timmo.systembridge.shared.Constants.CONNECTION_HOST
+import dev.timmo.systembridge.shared.Constants.CONNECTION_NAME
+import dev.timmo.systembridge.shared.Constants.CONNECTION_UID
+import dev.timmo.systembridge.shared.Constants.DEFAULT_API_PORT
+import dev.timmo.systembridge.shared.Constants.SETUP_EDIT
 import dev.timmo.systembridge.R
 import dev.timmo.systembridge.data.*
+import dev.timmo.systembridge.data.bridge.Endpoints
+import dev.timmo.systembridge.data.bridge.Information
+import dev.timmo.systembridge.shared.ServiceBuilder
 import kotlinx.coroutines.DelicateCoroutinesApi
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.GlobalScope
