+++ app/src/main/java/dev/timmo/systembridge/activity/SettingsActivity.kt
@@ -14,12 +14,12 @@ import android.widget.TextView
 import androidx.appcompat.app.AppCompatActivity
 import androidx.recyclerview.widget.GridLayoutManager
 import androidx.recyclerview.widget.RecyclerView
-import dev.timmo.systembridge.Constants.CONNECTION_API_KEY
-import dev.timmo.systembridge.Constants.CONNECTION_API_PORT
-import dev.timmo.systembridge.Constants.CONNECTION_HOST
-import dev.timmo.systembridge.Constants.CONNECTION_NAME
-import dev.timmo.systembridge.Constants.CONNECTION_UID
-import dev.timmo.systembridge.Constants.SETUP_EDIT
+import dev.timmo.systembridge.shared.Constants.CONNECTION_API_KEY
+import dev.timmo.systembridge.shared.Constants.CONNECTION_API_PORT
+import dev.timmo.systembridge.shared.Constants.CONNECTION_HOST
+import dev.timmo.systembridge.shared.Constants.CONNECTION_NAME
+import dev.timmo.systembridge.shared.Constants.CONNECTION_UID
+import dev.timmo.systembridge.shared.Constants.SETUP_EDIT
 import dev.timmo.systembridge.R
 import dev.timmo.systembridge.data.AppDatabase
 import dev.timmo.systembridge.data.Connection
