+++ app/src/main/java/dev/timmo/systembridge/data/Connection.kt
@@ -3,11 +3,11 @@ package dev.timmo.systembridge.data
 import androidx.room.ColumnInfo
 import androidx.room.Entity
 import androidx.room.PrimaryKey
-import dev.timmo.systembridge.Constants.CONNECTION_API_KEY
-import dev.timmo.systembridge.Constants.CONNECTION_API_PORT
-import dev.timmo.systembridge.Constants.CONNECTION_HOST
-import dev.timmo.systembridge.Constants.CONNECTION_NAME
-import dev.timmo.systembridge.Constants.CONNECTION_UUID
+import dev.timmo.systembridge.shared.Constants.CONNECTION_API_KEY
+import dev.timmo.systembridge.shared.Constants.CONNECTION_API_PORT
+import dev.timmo.systembridge.shared.Constants.CONNECTION_HOST
+import dev.timmo.systembridge.shared.Constants.CONNECTION_NAME
+import dev.timmo.systembridge.shared.Constants.CONNECTION_UUID
 
 @Entity
 data class Connection(
