--- app/src/main/java/dev/timmo/systembridge/data/Open.kt
@@ -1,5 +1,0 @@
-package dev.timmo.systembridge.data
-
-data class Open(
-    val path: String,
-)
