+++ app/src/main/java/dev/timmo/systembridge/data/bridge/Open.kt
@@ -1,0 +1,5 @@
+package dev.timmo.systembridge.data.bridge
+
+data class Open(
+    val path: String,
+)
